introduce target service to get destinations
