+++ pkg/config/playbook.go
@@ -50,30 +50,12 @@ type SimplePlayBook struct {
 	Task      []Cmd    `yaml:"task" toml:"task"`           // single task is a list of commands
 }
 
-// Target defines hosts to run commands on
-type Target struct {
-	Name   string        `yaml:"name" toml:"name"`     // name of target
-	Hosts  []Destination `yaml:"hosts" toml:"hosts"`   // direct list of hosts to run commands on, no need to use inventory
-	Groups []string      `yaml:"groups" toml:"groups"` // list of groups to run commands on, matches to inventory
-	Names  []string      `yaml:"names" toml:"names"`   // list of host names to run commands on, matches to inventory
-	Tags   []string      `yaml:"tags" toml:"tags"`     // list of tags to run commands on, matches to inventory
-}
-
 // Task defines multiple commands runs together
 type Task struct {
 	Name     string `yaml:"name" toml:"name"` // name of target, set by config caller
 	User     string `yaml:"user" toml:"user"`
 	Commands []Cmd  `yaml:"commands" toml:"commands"`
 	OnError  string `yaml:"on_error" toml:"on_error"`
-}
-
-// Destination defines destination info
-type Destination struct {
-	Name string   `yaml:"name" toml:"name"`
-	Host string   `yaml:"host" toml:"host"`
-	Port int      `yaml:"port" toml:"port"`
-	User string   `yaml:"user" toml:"user"`
-	Tags []string `yaml:"tags" toml:"tags"`
 }
 
 // Overrides defines override for task passed from cli
@@ -297,17 +279,6 @@ func (p *PlayBook) Task(name string) (*Task, error) {
 	return res, nil
 }
 
-// AllSecretValues returns all secret values from all tasks and all commands.
-// It is used to mask Secrets in logs.
-func (p *PlayBook) AllSecretValues() []string {
-	res := make([]string, 0, len(p.secrets))
-	for _, v := range p.secrets {
-		res = append(res, v)
-	}
-	sort.Strings(res)
-	return res
-}
-
 // TargetHosts returns target hosts for given target name.
 // After it gets destinations from targetHosts(name) it applies overrides of user, set default port 22 if needed
 // and deduplicate results.
@@ -339,7 +310,8 @@ func (p *PlayBook) TargetHosts(name string) ([]Destination, error) {
 		return p.User
 	}
 
-	res, err := p.targetHosts(name)
+	tgSvc := newTargetService(p.Targets, p.User, p.inventory)
+	res, err := tgSvc.destinations(name)
 	if err != nil {
 		return nil, err
 	}
@@ -355,138 +327,15 @@ func (p *PlayBook) TargetHosts(name string) ([]Destination, error) {
 	return dedup(res), nil
 }
 
-// targetHosts returns target hosts for given target name.
-// The result is not deduplicated and not modified with overrides.
-func (p *PlayBook) targetHosts(name string) ([]Destination, error) {
-	t, ok := p.Targets[name] // get target from playbook
-	if ok {
-		if len(t.Hosts) == 0 && len(t.Names) == 0 && len(t.Groups) == 0 && len(t.Tags) == 0 {
-			return nil, fmt.Errorf("target %q has no hosts, names, tags or groups", name)
-		}
-		log.Printf("[DEBUG] target %q found in playbook", name)
-		// we have found target in playbook, process hosts, names and group
-		res := []Destination{}
-
-		if len(t.Hosts) > 0 {
-			// target has "hosts", use all of them as is
-			res = append(res, t.Hosts...)
-			log.Printf("[DEBUG] target %q has %d hosts: %+v", name, len(t.Hosts), t.Hosts)
-		}
-
-		if len(t.Names) > 0 && p.inventory != nil {
-			// target has "names", match them to "all" group in inventory by name
-			for _, n := range t.Names {
-				for _, h := range p.inventory.Groups[allHostsGrp] {
-					if strings.EqualFold(h.Name, n) {
-						res = append(res, h)
-						log.Printf("[DEBUG] target %q found name match %+v", name, h)
-						break
-					}
-				}
-			}
-		}
-
-		if len(t.Groups) > 0 && p.inventory != nil {
-			// target has "groups", get all hosts from inventory for each group
-			for _, g := range t.Groups {
-				// we don't set default port and user here, as they are set in inventory already
-				res = append(res, p.inventory.Groups[g]...)
-				log.Printf("[DEBUG] target %q found group match %+v", name, p.inventory.Groups[g])
-			}
-		}
-
-		if len(t.Tags) > 0 && p.inventory != nil {
-			// target has "tags", get all hosts from inventory for each tag
-			for _, tag := range t.Tags {
-				for _, h := range p.inventory.Groups[allHostsGrp] {
-					if len(h.Tags) == 0 {
-						continue
-					}
-					for _, t := range h.Tags {
-						if strings.EqualFold(t, tag) {
-							res = append(res, h)
-							log.Printf("[DEBUG] target %q found tag match %+v", name, h)
-						}
-					}
-				}
-			}
-		}
-
-		if len(res) == 0 {
-			return nil, fmt.Errorf("hosts for target %q not found", name)
-		}
-		log.Printf("[DEBUG] target %q has %d total hosts: %+v", name, len(res), res)
-		return res, nil
+// AllSecretValues returns all secret values from all tasks and all commands.
+// It is used to mask Secrets in logs.
+func (p *PlayBook) AllSecretValues() []string {
+	res := make([]string, 0, len(p.secrets))
+	for _, v := range p.secrets {
+		res = append(res, v)
 	}
-
-	// target not defined in playbook
-	log.Printf("[DEBUG] target %q not found in playbook", name)
-
-	// try first as group in inventory
-	hosts, ok := p.inventory.Groups[name]
-	if ok {
-		res := make([]Destination, len(hosts))
-		copy(res, hosts)
-		log.Printf("[DEBUG] target %q found as group in inventory: %+v", name, res)
-		return res, nil
-	}
-
-	// try as a tag in inventory
-	res := []Destination{}
-	for _, h := range p.inventory.Groups[allHostsGrp] {
-		if len(h.Tags) == 0 {
-			continue
-		}
-		for _, t := range h.Tags {
-			if strings.EqualFold(t, name) {
-				res = append(res, h)
-			}
-		}
-	}
-	if len(res) > 0 {
-		log.Printf("[DEBUG] target %q found as tag in inventory: %+v", name, res)
-		return res, nil
-	}
-
-	// try as single host name in inventory
-	for _, h := range p.inventory.Groups[allHostsGrp] {
-		if strings.EqualFold(h.Name, name) {
-			log.Printf("[DEBUG] target %q found as name in inventory: %+v", name, h)
-			return []Destination{h}, nil
-		}
-	}
-
-	// try as a single host address in inventory
-	for _, h := range p.inventory.Groups[allHostsGrp] {
-		if strings.EqualFold(h.Host, name) {
-			log.Printf("[DEBUG] target %q found as host in inventory: %+v", name, h)
-			return []Destination{h}, nil
-		}
-	}
-
-	// try as single host or host:port or user@host:port
-	user := p.User
-	if strings.Contains(name, "@") { // extract user from name
-		elems := strings.Split(name, "@")
-		user = elems[0]
-		if len(elems) > 1 {
-			name = elems[1] // skip user part
-		}
-	}
-
-	if strings.Contains(name, ":") {
-		elems := strings.Split(name, ":")
-		port, err := strconv.Atoi(elems[1])
-		if err != nil {
-			return nil, fmt.Errorf("can't parse port %s: %w", elems[1], err)
-		}
-		log.Printf("[DEBUG] target %q used as host:port %s:%d", name, elems[0], port)
-		return []Destination{{Host: elems[0], Port: port, User: user}}, nil
-	}
-
-	// finally we assume it is a host name, with default port 22
-	log.Printf("[DEBUG] target %q used as host:22 %s", name, name)
-	return []Destination{{Host: name, Port: 22, User: user}}, nil
+	sort.Strings(res)
+	return res
 }
 
 // loadInventoryFile loads inventory from file or url and returns a struct with groups.
